// src/types/MapData.type.ts
// Generated by https://quicktype.io

export interface MapList {
    id:         string;
    seed:       number;
    difficulty: number;
    maps:       { [key: string]: MapData };
}

export interface MapData {
    type:    string;
    id:      number;
    name:    string;
    offset:  Offset;
    size:    Size;
    objects: Object[];
    map:     Array<number[]>;
}

export interface Object {
    id:    number;
    type:  Type;
    x:     number;
    y:     number;
    name?: string;
}

export enum Type {
    Exit = "exit",
    NPC = "npc",
    Object = "object",
}

export interface Offset {
    x: number;
    y: number;
}

export interface Size {
    width:  number;
    height: number;
}